Refactor: Update RAM calculation to use Long for better precision
This commit updates the return type of the RAM usage utility to prevent potential integer overflow during memory calculations.

**Refactoring & Bug Fixes:**
- **Storage Utilities (`StorageUtils.kt`)**: Updated `getUsedRAM()` to return a `Long` instead of an `Int`, removing the explicit integer cast to ensure more accurate memory reporting.

+++ app/src/main/java/com/lkonlesoft/displayinfo/utils/StorageUtils.kt
@@ -73,9 +73,9 @@ class StorageUtils (private val context: Context) {
         return info.availMem / 1024 / 1024 // MB
     }
 
-    fun getUsedRAM(): Int {
+    fun getUsedRAM(): Long {
         am.getMemoryInfo(info)
-        return (info.totalMem - info.availMem).toInt() / 1024 / 1024 // MB
+        return (info.totalMem - info.availMem) / 1024 / 1024 // MB
     }
 
     private fun getStorageStats(path: File): Pair<Long, Long> {
